fix coverity CID 49579701
CID 49579701: (#4 of 4): CERT-C Integers (CERT INT31-C)
1. cert_int31_c_violation: Casting irq from enum to unsigned int
without checking its value may result in lost or misinterpreted data.

+++ apps/hello_world/freertos/boards/imx943evk/ca55/hardware_init.c
@@ -26,7 +26,7 @@ void BOARD_InitHardware(void)
 	BOARD_InitMemory();
 	/* Enable GIC before register any interrupt handler*/
 	GIC_Enable();
-	os_irq_register(irq, SM_platform_MU_IRQHandler, NULL, portLOWEST_USABLE_INTERRUPT_PRIORITY - 1);
+	os_irq_register((unsigned int)irq, SM_platform_MU_IRQHandler, NULL, portLOWEST_USABLE_INTERRUPT_PRIORITY - 1);
 	SM_Platform_Init();
 	board_pins_setup();
 	board_clock_setup();
