+++ apps/rpmsg_str_echo/freertos/boards/imx95lpd5evk19/ca55/app_board.c
@@ -18,7 +18,7 @@ void board_init(void)
 	BOARD_InitMemory();
 	/* Enable GIC before register any interrupt handler*/
 	GIC_Enable();
-	os_irq_register(irq, SM_platform_MU_IRQHandler, NULL, portLOWEST_USABLE_INTERRUPT_PRIORITY - 1);
+	os_irq_register((unsigned int)irq, SM_platform_MU_IRQHandler, NULL, portLOWEST_USABLE_INTERRUPT_PRIORITY - 1);
 	SM_Platform_Init();
 	board_pins_setup();
 	board_clock_setup();
