+++ apps/rt_latency/freertos/boards/imx95lpd5evk19/ca55/rt_latency_board.c
@@ -26,7 +26,7 @@ void rt_latency_board_init(void)
 	/* Enable GIC before register any interrupt handler*/
 	GIC_Enable();
 	SM_Platform_Init();
-	os_irq_register(irq, SM_platform_MU_IRQHandler, NULL, portLOWEST_USABLE_INTERRUPT_PRIORITY - 1);
+	os_irq_register((unsigned int)irq, SM_platform_MU_IRQHandler, NULL, portLOWEST_USABLE_INTERRUPT_PRIORITY - 1);
 	board_pins_setup();
 	board_clock_setup();
 	BOARD_InitDebugConsole();
